requirements.txt
streamlit
pandas
plotly
asyncio
python-dotenv
aiohttp
requests
python-docx
playwright
beautifulsoup4